New feature: Sound packs
The default notification and call sounds in Element have long been
detested. This feature allows for many custom sound packs and the user
can switch between them at will.

The new default sound pack is Schildi. People can switch back to
Element's sounds if they prefer.

The new ring sound (and ringback) is copyright (c) Ana Gelez, Apache
License, and has been offered as a sound for Element, though was never
accepted. The new message notification sound is public domain.

Sources:
https://github.com/matrix-org/matrix-react-sdk/pull/4500/files
https://heyquark.com/general/2022/07/04/microblogging-2/

A desirable feature would be some way to preview the selected sounds
in the settings menu, but that's not a showstopper.