Fix custom emotes height attribute in URL
As discussed in the Schildi development room.

The current behaviour happens to work well for people with dynamic thumbnails disabled in Synapse (the default) but a good implementation would provide images that are far smaller than requested, making them look blurry in the timeline, especially when enlarged in an emoji-only message.

This makes the URL use 18*DPI for normal custom emojis and 48*DPI for big custom emojis, which should be the perfect size for display.

GIF emojis still don't work. Could be a problem for Synapse to solve.