Create a dedicated CSS file for customization, fixing line height on IRC layout
This commit does two things:

1. Create a dedicated CSS for customization, making it modular based on mixin
2. Fix line-height on IRC layout

When it comes to styling, basically we could just import a custom CSS file after the last existing `@import` on _components.pcss and put every custom stuff like CSS declarations and variables in it to let them override the inherited values in the way which the concept of CSS expects. This commit also takes advantage of PostCSS to make our style codebase modular by creating files, on which only mixins are defined to be used elsewhere.

The obvious merit of doing so is that it will remove the burden of applying customization directly on the upstream CSS codebase fixing conflicts. Separating customized styles from the upstream codebase should also greatly reduce the manpower to rebase, without being worried about possible regressions. Doing so will also make it easier to fix the regressions on our codebase which the upstream project has not fixed yet.

Even if the properties are annoyingly styled with `!important` by the upstream to cover the regression, they could be overridden by ones with `!important` on the CSS files imported later than that.

The way in which the upstream project generates concatenated CSS files has been very stable (essentially same since at least 2018. See: https://github.com/matrix-org/matrix-react-sdk/commits/develop/res/css/rethemendex.sh), so we should be able to depend on the current way how it works for a reasonable time.